refactor(shim): PR-9.4 — 删 project_jobs / projects / versions / versions_phase 4 shim
batch 2 of 6 — projects 域 4 个 top-level shim。最大头（90+ caller）。

  删 studio/{project_jobs,projects,versions,versions_phase}.py（4 × 10 行）

caller 更新（~50 文件，Python 脚本批量重写 + 手工补漏）：
- `from studio import db, project_jobs, projects, versions` 拆成两行
- `from studio.versions import ...` → `from studio.services.projects.versions import ...`
- `from studio.versions_phase import ...` → `from studio.services.projects.phase import ...`
- 内部 `from .... import projects, versions, ...` 拆出 services.projects 入
- 4 个手工补漏：
  - studio/services/version_config.py: from ..versions → from .projects.versions
  - studio/services/dataset/curation.py: from ... import projects, versions → ..projects
  - tests/test_versions_reconcile.py + test_version_status_phase.py: 拆 import 行

回归：1588 passed (2 pre-existing imagehash fail)。
