+++ studio/api/routers/projects/_shared.py
@@ -10,7 +10,8 @@ from typing import Any
 
 from fastapi import HTTPException
 
-from .... import db, projects, versions
+from .... import db
+from ....services.projects import projects, versions
 from ....event_bus import bus
 from ....services import version_config
