+++ studio/api/routers/projects/exports.py
@@ -28,7 +28,8 @@ from ._shared import (
     _publish_project_state,
     _publish_version_state,
 )
-from .... import db, projects, versions
+from .... import db
+from ....services.projects import projects, versions
 from ....event_bus import bus
 from ....paths import DATA_EXPORTS, REPO_ROOT, USER_PRESETS_DIR
 from ....services import train_io
