+++ studio/api/routers/queue/outputs.py
@@ -24,7 +24,8 @@ from fastapi.responses import FileResponse
 
 from ...errors import _export_result, _safe_join_or_400, _unique_data_export_path
 from ...schemas.queue import ExportOutputsBody
-from .... import db, projects, versions
+from .... import db
+from ....services.projects import projects, versions
 from ....event_bus import bus
 from ....paths import DATA_EXPORTS
