+++ studio/services/preprocess/core.py
@@ -32,7 +32,7 @@ from __future__ import annotations
 from pathlib import Path
 from typing import Any, Iterable, Optional
 
-from ... import project_jobs, projects
+from ...services.projects import jobs as project_jobs, projects
 from ...services.dataset.scan import IMAGE_EXTS
 from . import manifest as preprocess_manifest
