+++ tests/test_phase_endpoints.py
@@ -7,7 +7,8 @@ from pathlib import Path
 import pytest
 from fastapi.testclient import TestClient
 
-from studio import db, project_jobs, projects, secrets, server, versions
+from studio import db, secrets, server
+from studio.services.projects import jobs as project_jobs, projects, versions
 
 
 @pytest.fixture
