+++ tests/test_preprocess_crop_worker.py
@@ -11,7 +11,7 @@ from pathlib import Path
 import pytest
 from PIL import Image
 
-from studio import projects
+from studio.services.projects import projects
 from studio.services import preprocess_manifest as pm
 from studio.workers import preprocess_worker as worker
