+++ tests/test_projects.py
@@ -5,7 +5,8 @@ from pathlib import Path
 
 import pytest
 
-from studio import db, projects
+from studio import db
+from studio.services.projects import projects
 
 
 @pytest.fixture
