+++ tests/test_train_io.py
@@ -8,7 +8,8 @@ from pathlib import Path
 
 import pytest
 
-from studio import db, projects, versions
+from studio import db
+from studio.services.projects import projects, versions
 from studio.services import train_io
