+++ tests/test_versions.py
@@ -5,7 +5,8 @@ from pathlib import Path
 
 import pytest
 
-from studio import db, projects, versions
+from studio import db
+from studio.services.projects import projects, versions
 
 
 @pytest.fixture
