+++ tests/test_versions_reconcile.py
@@ -6,7 +6,8 @@ from pathlib import Path
 
 import pytest
 
-from studio import db, projects, versions
+from studio import db
+from studio.services.projects import projects, versions
 
 
 @pytest.fixture
